requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pypi_0
argon2-cffi=21.1.0=pypi_0
argparse=1.4.0=pypi_0
async-generator=1.10=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=py_0
blas=1.0=mkl
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py39h06a4308_0
cffi=1.14.6=pypi_0
charset-normalizer=2.0.5=pypi_0
click=8.0.3=pypi_0
cloudpickle=2.0.0=pyhd3eb1b0_0
configparser=5.0.2=pypi_0
construct=2.10.54=pypi_0
cryptography=3.4.8=pypi_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=py39h06a4308_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2021.10.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.4.1=py39h295c915_0
decorator=4.4.2=py_0
deepspeed=0.5.4=pypi_0
docker-pycreds=0.4.0=pypi_0
expat=2.4.1=h2531618_2
ffmpeg=4.3=hf484d3e_0
filelock=3.3.0=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2022.1.0=pyhd3eb1b0_0
future=0.18.2=pypi_0
gdrive=0.0.8=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.24=pypi_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
jsonpatch=1.32=pypi_0
jsonpointer=2.2=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py39h06a4308_0
kiwisolver=1.3.1=py39h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
locket=0.2.1=py39h06a4308_1
lxml=4.6.3=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
matplotlib=3.4.2=py39h06a4308_0
matplotlib-base=3.4.2=py39hab158f2_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.0=py39h42c9631_2
mkl_random=1.2.2=py39h51133e4_0
munkres=1.1.4=py_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2.2=pypi_0
numpy=1.20.3=py39hf144106_0
numpy-base=1.20.3=py39h74d4b33_0
oauth2client=4.1.3=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.3.56=pypi_0
openh264=2.1.0=hd408876_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1m=h7f8727e_0
outcome=1.1.0=pypi_0
packaging=21.0=pypi_0
parso=0.8.0=py_0
partd=1.2.0=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=py39h2c7a002_0
pip=21.2.4=py37h06a4308_0
progressbar=2.5=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.8=py_0
protobuf=3.17.3=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pycryptodomex=3.10.1=pypi_0
pydrive=1.3.1=pypi_0
pygments=2.7.1=py_0
pykeepass=4.0.1=pypi_0
pyopenssl=21.0.0=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyperclip=1.8.2=pypi_0
pyqt=5.9.2=py39h2531618_6
pysimplegui=4.47.0=pypi_0
python=3.9.6=h12debd9_1
python-dateutil=2.8.1=py_0
pytorch=1.9.0=py3.9_cuda11.1_cudnn8.0.5_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=py39h6323ea4_4
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py39h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.16.2=py39ha9443f7_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.1=py39h292c36d_2
secretstorage=3.2.0=pypi_0
selenium=4.1.0=pypi_0
sentry-sdk=1.4.3=pypi_0
setuptools=56.0.0=pypi_0
shortuuid=1.0.1=pypi_0
sip=4.19.13=py39h2531618_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.36.0=hc218d9a_0
subprocess32=3.5.4=pypi_0
tabulate=0.8.9=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=1.8=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.10=hbc83047_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.9.0=py39
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
torchvision=0.10.0=py39_cu111
tornado=6.1=py39h27cfd23_0
tqdm=4.62.2=pypi_0
traitlets=5.0.5=py_0
trio=0.19.0=pypi_0
trio-websocket=0.9.2=pypi_0
triton=1.1.0=pypi_0
typing_extensions=3.10.0.0=pyhca03da5_0
tzdata=2021a=h5d7bf9c_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.6=pypi_0
versioneer=0.20=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.2.5=py_0
websocket-client=1.2.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.32.3=pypi_0
wsproto=1.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yaspin=2.1.0=pypi_0
zeromq=4.3.3=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
